Refactored function definition
Now it uses a function/factory instead of a weird array. In the future this could be changed to allow some kind of inheritance of method calls... but that's probably overkill considering each function is technically part of some API (which in the cases of Blogger, MetaWeblog are never going to change...) so hopefully I'll leave it as is.